# Hostlify.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hostlify.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/History/GetAll": {
      "get": {
        "tags": [
          "History"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/History/byManagerId": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "managerId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/History/byResource": {
      "post": {
        "tags": [
          "History"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HistoryResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/History/{id}": {
      "delete": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/GetAll": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/byManagerId": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "managerId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/byGuestId": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "guestId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/byRoomName": {
      "get": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roomName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoomResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/Rooms/register/{id}": {
      "put": {
        "tags": [
          "Rooms"
        ],
        "operationId": "register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterGuestResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/{id}": {
      "put": {
        "tags": [
          "Rooms"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateRoomRegisteredResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Rooms/evict/{id}": {
      "put": {
        "tags": [
          "Rooms"
        ],
        "operationId": "Evict",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/GetAllFoodServices": {
      "get": {
        "tags": [
          "Services"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/GetAllCleaningServices": {
      "get": {
        "tags": [
          "Services"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/FoodServicebyRoomId": {
      "get": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roomId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/CleaningServicebyRoomId": {
      "get": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roomId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/FoodServiceByResource": {
      "post": {
        "tags": [
          "Services"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FoodServicesResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/CleaningServiceByResource": {
      "post": {
        "tags": [
          "Services"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PostCleaningServiceResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/deleteFoodService/{id}": {
      "delete": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Services/deleteCleaningService/{id}": {
      "delete": {
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/LoginUserResponse"
            }
          }
        }
      }
    },
    "/api/User/Signup": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/User/GetByEmail/{email}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetByEmail",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/User/GetRoomsLimit/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/User/PostRoomsLimit": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/postRoomLimitResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "FoodServicesResource": {
      "required": [
        "roomId"
      ],
      "type": "object",
      "properties": {
        "roomId": {
          "format": "int32",
          "type": "integer"
        },
        "dish": {
          "type": "string"
        },
        "dishQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "drink": {
          "type": "string"
        },
        "drinkQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "cream": {
          "type": "string"
        },
        "creamQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "instruction": {
          "type": "string"
        }
      }
    },
    "HistoryResource": {
      "required": [
        "roomName"
      ],
      "type": "object",
      "properties": {
        "roomName": {
          "type": "string"
        },
        "managerId": {
          "format": "int32",
          "type": "integer"
        },
        "guestName": {
          "type": "string"
        },
        "initialDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "LoginResource": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "LoginUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "plan": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "jwt": {
          "type": "string"
        }
      }
    },
    "PostCleaningServiceResource": {
      "type": "object",
      "properties": {
        "roomId": {
          "format": "int32",
          "type": "integer"
        },
        "instruction": {
          "type": "string"
        }
      }
    },
    "postRoomLimitResource": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "roomsLimit": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "RegisterGuestResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "initialDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "RoomResource": {
      "type": "object",
      "properties": {
        "roomName": {
          "maxLength": 15,
          "type": "string"
        },
        "managerId": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "UpdateRoomRegisteredResource": {
      "type": "object",
      "properties": {
        "roomName": {
          "type": "string"
        },
        "guestId": {
          "format": "int32",
          "type": "integer"
        },
        "managerId": {
          "format": "int32",
          "type": "integer"
        },
        "initialDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "emergency": {
          "type": "boolean"
        },
        "servicePending": {
          "type": "boolean"
        }
      }
    },
    "UserResource": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "plan": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": { }
  },
  "security": [
    {
      "bearerAuth": [ ]
    }
  ]
}